A few more mediocre animations
Idle B is a basic idle pose, with just some slight motion. Idle A is
supposed to be a bit of a stretch, but it needs at least one additional
frame in the middle (which as of yet I can't add without redoing the
rest of it). Yawn really needs to be slowed down at the apex, which
could be done with additional frames but I think it would be good to be
able to specify delays.