Make mouse use client positions again
Fix panning momentum bugginess due to mouse position being updated only
on mousemove and not when the view had moved.

It wasn't really a good idea to specialize Mouse like that, changing the
semantics of its properties x and y; and I didn't think about updating
when the view moved at the time.