Include player's velocity in arrow velocity
Firing an arrow straight upwards while running, the arrow should roughly
follow your horizontal movement, and, on a level plane, hit near you.

As for the angle of the arrow, I'm not sure what it should look like,
but this probably isn't right. It's rotating as soon as it's fired.