Apply some drag to the fletching. I think it's working.
It looks PLAUSIBLE. I don't know a lot about archery, and I would really
love to see some scientific data on the flight paths of arrows.

The arrows are now once again angling forwards, and now in a way that's
defined in terms of a drag force. That seems kinda legit. There'd also
be a drag force on the tip and shaft, which I'm not modeling, and also
the tip would have a greater mass than the fletching, right? I don't
know how significant these things are.