Fix gamepad polling in Chrome
https://developer.mozilla.org/en-US/docs/Web/API/Gamepad_API/Using_the_Gamepad_API

"Behind the scenes, this object will be updated every time the gamepad's state changes."
It seems like not?

...Later on:
"Chrome does things differently here. Instead of constantly storing the gamepad's latest state in a variable it only stores a snapshot, so to do the same thing in Chrome you have to keep polling it"
Ugh! Come on!
First you give a complicated example when a simple one will do, then you give a contrived example of how to use the simpler API but involve all the complexity of the event based API, and then you describe this property of updating that you can rely on only in Firefox without mentioning that it's not cross-browser, and then later you note that it's not cross-browser, after I've tried to copy from the example and adapt it to my codebase.
MDN, I swear, sometimes I oughta shake my fist and say "Why I oughta..."
No, MDN is generally good though.