Optimize (for now at least)
I may want to continue to look for more targets to lodge into later,
to support, for instance, pinning an enemy by skewering its foot to the
ground. For now, this is helping mainly to avoid too many console logs
which can be confusing and which Chrome handles very poorly.