Project point outside terrain instead of backing up
When backing up, the particle was prone to becoming fixed in place,
stuck inside of the terrain. It only happened when shooting the arrow
in certain trajectories, but once it was in, the "backing up" strategy
(which I knew was a bad strategy before I coded it) has no way of
bringing it outside the terrain. (Modifying it to back up further than
its velocity doesn't help because it doesn't know which way to go,
it just adds energy to the system and causes weirder behavior.)
The only robust way to do it is to project the point outside the ground,
so there is no "stuck" state. Except for self-intersecting terrain,
I suppose... Crap, I have to test this...
...or two pieces of terrain that overlap. Crap.