Recompute velocity after projection onto terrain surface
I tried this before making the COR apply only perpendicularly, and it
didn't help, but trying it now, it's reduced the very aphysical behavior
introduced in the previous commit, but it's still weirdly hopping off of
the ground as it slides down it. And it still stops instantly sometimes,
even though I haven't tried to implement explicit friction yet.