Fix ugly cloud uniformity by using simplex noise as intended
The tiamblia-original implementation used simplex noise for the clouds,
albeit with a different library, and I just hacked in some sine waves
to get it working at a basic level. It was failing to generate unique
clouds due to seedrandom*, so I couldn't see how bad it was.
(*fixed in 383164eccda68d23242841b4aea7f67e24a4c8f7)