Refactor entity class checking to use class name strings
For maintainability, I want to move the entity sorting code out into a
new file. I also don't want to maintain require statements for all the
classes I'm checking against. To that end, I'm disentangling the
require statements used for adding entities to the registry from the
instance-of checking.

As a downside, maintenance-wise, I'm now relying on strings to match
class names, in a way that can't be simply statically analyzed, and I
can't Ctrl+click to go to the definition of a class specified by string.

It's also more code. And probably slower.

This isn't worth it, I'm going to revert this.