Prevent clicking welcome text, as it interferes with bow action
The welcome message doesn't have a well defined boundary, so it's extra
annoying when clicking doesn't work.

Normally I try to preserve the ability to select text, but it might not
be worth it in this case. I could however make it so aiming and shooting
the bow and arrow works when clicking outside the canvas...