Smooth out kinks by checking from current position
This allows the segment distance constraints to do their work.
It also slows the crawl speed down a lot because instead of moving at
essentially a fixed rate, it's dampened by the fixity of the attachment
constraint.

This doesn't handle kinks where the points fully double up, as the
distance constraints can be fully satisfied in that case.