WIP: smooth out normals so caterpillar can pick a side
I was hoping this would help the caterpillar stay on one side of a tree branch instead of oscillating between sides.
I also wanted this to help with the case where part of the caterpillar doesn't contact the ground and so no leg is shown for it, because that leg doesn't know where to go. I wanted it to "share the knowledge", so to speak, between the sections of the bugger.
But this isn't working. It's not the right solution. Garbage in, garbage out. At the point that this code operates, it's already flip flopping between sides of a branch of a tree.
The problem is that the way it's determining the "towards ground" or "away from ground" vectors just doesn't work for thin lines. I wrote it with polygons in mind, and it only works for polygons.
(Side note: using point.smoothed_normal in the "search towards the ground, in the direction it was last found" code doesn't help either.)