Refactor: flip normal
This makes two negatives positive, and it makes two positives negative,
but the two new negatives are going to go away. One is just a debug viz,
I can just flip it back; I'm just keeping this commit a clean refactor.
For the other, I plan on making the foot attach to the ground instead of
having an offset from the body part, so I might do away with it.