Tweak cactus shadow/highlight
I'm mainly trying to make this look less "mobile game" like.
Mobile games often have a very glossy look, and I'm going for a more
flat look. That said, I'm not flattening this out, as such, at least not
completely, because the shadows and highlights really help to sell it
as a cactus.

This makes it more realistic, with the ridge highlights that are within
the branch shadow being more subdued.