Fix elbow bending backwards due to shoulder vs sternum height difference
This also makes it not bend the arm all the way, since it's still using
the sternum to determine the arm extent, even though it could reach
further from the shoulder than from the sternum, at least in some cases.
In other cases I suppose it would be the opposite, and it might actually
overextend.