Divide terrain in the middle of buckets so it fits in 2 instead of 3
Make the cut in the middle of each bucket, so that the resulting terrain
polygons fit within two buckets. Before, they landed in three buckets
despite being barely wider than one!

As a side effect, this makes one of the caterpillars in the world
start out in a seam, and it either crawls up or down through it,
randomly, since it starts out horizontal, and the seam is vertical,
and the caterpillar simulation is non-deterministic because it uses
performance.now()...
(I had placed the caterpillar inside the terrain so that it would align
itself to the terrain, without having to position it manually so close
to the contour of the polygon that each of its feet would touch.)