Remove Water exemption from Terrain optimization
It should no longer be a performance concern, since the reflection
effect will only be rendered once no matter how many polygons it's
split into. Oh, and the waves should work, they would've been a problem
as well if it was split up before, but now with the derived colliders
approach, splitting the polygon just for collision detection, it should
all be fine!