Reset player in world, mainly for the eyes
I dragged a new Player into the world, then copied over the x and y
in the new entity properties panel, as well as the id, and deleted
the old entity. I could also do this sort of thing by editing the world
file JSON directly, but it's fun to start making use of the lil-gui
entity properties inspector/editor.

None of these changes are terribly important, but I didn't like the
character staring dead-eyed at the camera in the initial state.