Deer: remove dir in favor of move_x
Of course this still lacks the behavior I had in tiamblia-original,
where the deer would have momentum. Right now it's stopping to quickly.
I might want a separate variable to implement behavior like that,
and other animalistic (and animistic) behaviors. But the animals are all
super dumb so far, so it's not like the old code will help much w/ that.

This unintentionally makes deer swim much faster. It was quite slow...