Don't show warning toast if Cancel clicked at permission prompt
For saving, it's a long message to read for a non-error scenario:
"Failed to save file: NotAllowedError: The request is not allowed by the user agent or the platform in the current context."

As for the case of clearing the auto-save, it's a little more complex,
we don't want to just leave it with an error message or no error message.
It's better with a bespoke message here.