Player: avoid transitory moment when aiming straight up (for head)
I want to pair this with changing the eyes/head to look like they're
looking upwards as well, but honestly, just avoiding that tweening
garbage code with garbage limitations, it feels worth it to remove.

There's still awkward tweening on the arms when flipping around,
and the arms and legs essentially switch identity, and start bending
the opposite directions when flipping. This would all be so much simpler
in 3D.